+++ R/splot.R
@@ -1,11 +1,11 @@
 #' Split Plot
 #'
 #' A plotting function aimed at automating some common visualization tasks in order to ease data exploration.
-#' @param y required. Primary variable, to be shown on the y axis unless \code{x} is not specified. Can be an object,
-#'   name of a column in data, or a formula (see note). Multiple variables can also be included in a matrix-like object.
+#' @param y A formula (see note), or the primary variable(s) to be shown on the y axis unless (\code{x} is not specified).
+#'   When not a formula, this can be one or more variables as objects or names in \code{data}.
 #' @param x secondary variable, to be shown in on the x axis. If not specified, \code{type} will be set to \code{'density'}.
 #'   If \code{x} is a factor or vector of characters, or has fewer than \code{lim} levels when treated as a factor,
-#'   \code{type} will be set to \code{"line"} unless specified.
+#'   \code{type} will be set to \code{'line'} unless specified.
 #' @param by the 'splitting' variable within each plot, by which the plotted values of \code{x} and \code{y} will be
 #'   grouped.
 #' @param between a single object or name, or two in a vector (e.g., \code{c(b1, b2)}), the levels of which will determine
@@ -25,11 +25,12 @@
 #' @param su a subset to all variables, applied after they are all retrieved from \code{data} or the environment.
 #' @param levels a list with entries corresponding to variable names, used to rename and/or reorder factor levels. To
 #'   reorder a factor, enter a vector of either numbers or existing level names in the new order (e.g.,
-#'   \code{levels=list(var=c(3,2,1))}). To rename levels of a factor, enter a character vector the same length as the number
-#'   of levels. To rename and reorder, enter a list, with names as the first entry, and order as the second entry (e.g.,
-#'   \code{levels=list(var=list(c('a','b','c'),c(3,2,1)))}). This happens after variables are split, so names and orders
-#'   should correspond to the new split levels of split variables. For example, if a continuous variable is median split,
-#'   it now has two levels ('Under Median' and 'Over Median'), which are the levels reordering or renaming would apply to.
+#'   \code{levels =}\code{list(var =} \code{c(3,2,1))}). To rename levels of a factor, enter a character vector the same
+#'   length as the number of levels. To rename and reorder, enter a list, with names as the first entry, and order as the
+#'   second entry (e.g., \code{levels =} \code{list(var =} \code{list(c('a','b','c'),} \code{c(3,2,1)))}). This happens after
+#'   variables are split, so names and orders should correspond to the new split levels of split variables. For example, if
+#'   a continuous variable is median split, it now has two levels ('Under Median' and 'Over Median'), which are the levels
+#'   reordering or renaming would apply to.
 #' @param error string: sets the type of error bars to show in bar or line plots, or turns them off. If \code{FALSE}, no error
 #'   bars will be shown. Otherwise, the default is \code{"standard error"} (\code{'^s'}), with \code{"confidence intervals"}
 #'   (anything else) as an option.
@@ -44,12 +45,13 @@
 #'   (\code{'^sm|^sp|^in'}) will use \code{\link[stats]{smooth.spline}}. \code{'connected'} (\code{'^e|^co|^d'}) will draw
 #'   lines connecting all points, and \code{FALSE} will not draw any lines.
 #' @param ... passes additional arguments to \code{\link[graphics]{par}} or \code{\link[graphics]{legend}}. The
-#'   \code{\link[graphics]{par}} options \code{col, mfrow, oma, mar, mgp, font.main, cex.main, font.lab, tcl, pch, lwd,} and
-#'   \code{xpd} are all set within the function, but will be overwritten if they are included in the call. For example,
-#'   \code{col} sets font colors in this case (as opposed to \code{colors} which sets line and point colors). The default is
-#'   \code{'#303030'} for a nice dark grey, but maybe you want to lighten that up: \code{col='#606060'}. After arguments have
-#'   been applied to \code{\link[graphics]{par}}, if any have not been used and match a \code{\link[graphics]{legend}} argument,
-#'   these will be applied to \code{\link[graphics]{legend}}.
+#'   \code{\link[graphics]{par}} options \code{col}, \code{mfrow}, \code{oma}, \code{mar}, \code{mgp}, \code{font.main},
+#'   \code{cex.main}, \code{font.lab}, \code{tcl}, \code{pch}, \code{lwd}, and \code{xpd} are all set within the function,
+#'   but will be overwritten if they are included in the call. For example, \code{col} sets font colors in this case
+#'   (as opposed to \code{colors} which sets line and point colors). The default is \code{'#303030'} for a nice dark grey,
+#'   but maybe you want to lighten that up: \code{col='#606060'}. After arguments have been applied to
+#'   \code{\link[graphics]{par}}, if any have not been used and match a \code{\link[graphics]{legend}} argument, these will
+#'   be applied to \code{\link[graphics]{legend}}.
 #' @param line.type a character setting the style of line (e.g., with points at joints) to be drawn in line plots. Default
 #'   is \code{'b'} if \code{error} is \code{FALSE}, and \code{'l'} otherwise. See the \code{line} argument of
 #'   \code{\link[graphics]{plot.default}} for options. \code{line.type='c'} can look nice when there aren't a lot of
@@ -66,9 +68,9 @@
 #' @param file.name a string with the name of the file to be save (excluding the extension, as this is added depending on
 #'   \code{format}).
 #' @param colors sets a color theme or manually specifies colors. Default theme is \code{"pastel"}, with \code{"dark"} and
-#'   \code{"bright"} as options. If set to \code{"grey"}, or if \code{by} has more than 9 levels, a grey scale is calculated
-#'   using \code{\link[grDevices]{grey}}. See the \code{col} parameter in \code{\link[graphics]{par}} for acceptable manual
-#'   inputs.
+#'   \code{"bright"} as options; these are passed to \code{\link{splot.color}}. If set to \code{"grey"}, or if \code{by} has
+#'   more than 9 levels, a grey scale is calculated using \code{\link[grDevices]{grey}}. See the \code{col} parameter in
+#'   \code{\link[graphics]{par}} for acceptable manual inputs.
 #' @param myl sets the range of the y axis (\code{ylim} of \code{\link[graphics]{plot}} or \code{\link[graphics]{barplot}}).
 #'   If not specified, this will be calculated from the data.
 #' @param mxl sets the range of the x axis (\code{xlim} of \code{\link[graphics]{plot}}). If not specified, this will be
@@ -79,16 +81,16 @@
 #' @param bw sets the smoothing bandwidth when plotting densities. Default is \code{'nrd0'}. See
 #'   \code{\link[stats]{density}}.
 #' @param adj adjusts the smoothing of densities (\code{adj * bw}). See \code{\link[stats]{density}}.
-#' @param leg sets the legend inside or outside the plot frames (when a character matching \code{'^i'}, or a character matching
-#'   \code{'^o'} or a number respectively), or turns it off (when \code{FALSE}). When inside, a legend is drawn in each plot
-#'   frame. When outside, a single legend is drawn either to the right of all plot frames, or within an empty plot frame. By
-#'   default, this will be determined automatically, tending to set legends outside when there are multiple levels of
-#'   \code{between}. A number will try and set the legend in an empty frame within the grid of plot frames. If there are no empty
-#'   frames, the legend will just go to the side as if \code{leg='outside'}.
-#' @param lpos sets the position of the legend within its frame (whether inside or outside of the plot frames) based on keywords
-#'   (see \code{\link[graphics]{legend}}. By default, when the legend is outside, \code{lpos} is either \code{'right'} when the
-#'   legend is in a right-hand column, or \code{'center'} when in an empty plot frame. When the legend is inside and \code{lpos}
-#'   is not specified, the legend will be placed automatically based on the data.
+#' @param leg sets the legend inside or outside the plot frames (when a character matching \code{'^i'}, or a character
+#'   matching \code{'^o'} or a number respectively), or turns it off (when \code{FALSE}). When inside, a legend is drawn in
+#'   each plot frame. When outside, a single legend is drawn either to the right of all plot frames, or within an empty plot
+#'   frame. By default, this will be determined automatically, tending to set legends outside when there are multiple levels
+#'   of \code{between}. A number will try and set the legend in an empty frame within the grid of plot frames. If there are
+#'   no empty frames, the legend will just go to the side as if \code{leg='outside'}.
+#' @param lpos sets the position of the legend within its frame (whether inside or outside of the plot frames) based on
+#'   keywords (see \code{\link[graphics]{legend}}. By default, when the legend is outside, \code{lpos} is either
+#'   \code{'right'} when the legend is in a right-hand column, or \code{'center'} when in an empty plot frame. When the
+#'   legend is inside and \code{lpos} is not specified, the legend will be placed automatically based on the data.
 #' @param lvn legend variable name. Logical: if \code{FALSE}, the names of by and between variables will not be shown before
 #'   their level (e.g., for a sex variable with a "female" level, "sex: female" would become "female" in the legend or above
 #'   each plot window).
@@ -132,19 +134,21 @@
 #'   \code{TRUE} by default, meaning only levels with data will be presented, and the layout of \code{between} levels
 #'   will be minimized. \code{x} only applies to bar or line plots. \code{by} relates to levels presented in the legend.
 #'   If \code{bet} is \code{FALSE}, the layout of \code{between} variables will be strict, with levels of \code{between[1]}
-#'   as rows, and levels of \code{between[2]} as columns -- if there are no data at an intersection of levels, the corresponding
-#'   panel will be blank. See the input section.
+#'   as rows, and levels of \code{between[2]} as columns -- if there are no data at an intersection of levels, the
+#'   corresponding panel will be blank. See the input section.
 #' @param prat panel ratio, referring to the ratio between plot frames and the legend frame when the legend is out. A single
 #'   number will make all panels of equal width. A vector of two numbers will adjust the ratio between plot panels and the
 #'   legend panel (e.g., \code{prat=c(3,1)} makes all plot panels a relative width of 3, and the legend frame a relative
 #'   width of 1).
 #' @param model logical: if \code{TRUE}, the summary of an interaction model will be printed.
 #' @param options a list with named arguments, useful for setting temporary defaults if you plan on using some of the same
-#'   options for multiple plots (e.g., \code{opt=list(type='bar', colors='grey', bg='#999999'); splot(x~y, options=opt)}).
+#'   options for multiple plots (e.g., \code{opt =} \code{list(type = 'bar',} \code{colors = 'grey',} \code{bg = '#999999');}
+#'   \code{splot(x~y,} \code{options = opt)}).
 #'   use \code{\link[base]{quote}} to include options that are to be evaluated within the function (e.g.,
-#'   \code{opt=list(su=quote(y>0))}).
+#'   \code{opt =} \code{list(su =} \code{quote(y>0))}).
 #' @param add evaluated within the function. Useful for adding things like lines to a plot while the parameters are still
-#'   those set by the function (e.g., \code{add=abline(v=mean(x),xpd=FALSE)} for a vertical line at the mean of x).
+#'   those set by the function (e.g., \code{add =} \code{abline(v =} \code{mean(x),} \code{xpd = FALSE)} for a vertical line
+#'   at the mean of x).
 #'
 #' @return A list containing data and settings is invisibly returned, which might be useful to check for errors.
 #' Each of these objects can also be pulled from within \code{add}:
@@ -174,8 +178,8 @@
 #' Named vector arguments like \code{font}, \code{cex}, and \code{drop} can be set with a single value, positionally, or
 #' with names. If a single value is entered (e.g., \code{drop=FALSE}), this will be applied to each level (i.e.,
 #' \code{c(x=FALSE,by=FALSE,bet=FALSE)}). If more than one value is entered, these will be treated positionally (e.g.,
-#' \code{cex=c(2,1.2)} would be read as \code{c(title=2,leg=1.2,note=.7)}). If values are named, only named values will be set,
-#' with other defaults retained (e.g., \code{cex=c(note=1.2)} would be read as \code{c(title=1.5,leg=1,note=1.2)}).
+#' \code{cex=c(2,1.2)} would be read as \code{c(title=2,leg=1.2,note=.7)}). If values are named, only named values will be
+#' set, with other defaults retained (e.g., \code{cex=c(note=1.2)} would be read as \code{c(title=1.5,leg=1,note=1.2)}).
 #'
 #' @note
 #' \strong{x-axis levels text}
@@ -238,7 +242,7 @@
 #' #zooming in on one of the windows
 #' splot(y~x*by, data=dat, su=bet1==1&bet2==0)
 #'
-#' #compairing an adjusted lm prediction line with a loess line
+#' #comparing an adjusted lm prediction line with a loess line
 #' #this could also be entered as y ~ poly(x,3)
 #' splot(y~x+x^2+x^3, data=dat, su=bet1==1&bet2==0&by==1, add={
 #'   lines(x[order(x)], loess(y~x)$fitted[order(x)], lty=2)
@@ -260,13 +264,13 @@
 
 splot=function(y,x=NULL,by=NULL,between=NULL,cov=NULL,type='',split='median',data=NULL,su=NULL,levels=list(),
   error='standard',error.color='#585858',error.lwd=2,lim=9,lines=TRUE,...,line.type='l',mv.scale='none',mv.as.x=FALSE,
-  save=FALSE,format=cairo_pdf,dims=dev.size(),file.name='splot',colors=NULL,myl=NULL,mxl=NULL,autori=TRUE,xlas=0,ylas=1,
-  bw='nrd0',adj=2,leg='outside',lpos='auto',lvn=TRUE,title=TRUE,labx=TRUE,laby=TRUE,lty=TRUE,lwd=2,sub=TRUE,ndisp=TRUE,
-  note=TRUE,font=c(title=2,leg=1,note=3),cex=c(title=1.5,leg=1,note=.7),sud=TRUE,labels=TRUE,labels.filter='_|\\.',
-  labels.trim=20,points=TRUE,points.first=TRUE,byx=TRUE,drop=c(x=TRUE,by=TRUE,bet=TRUE),prat=c(1,1),model=FALSE,
-  options=NULL,add=NULL){
+  save=FALSE,format=cairo_pdf,dims=dev.size(),file.name='splot',colors='pastel',myl=NULL,mxl=NULL,autori=TRUE,xlas=0,
+  ylas=1,bw='nrd0',adj=2,leg='outside',lpos='auto',lvn=TRUE,title=TRUE,labx=TRUE,laby=TRUE,lty=TRUE,lwd=2,sub=TRUE,
+  ndisp=TRUE,note=TRUE,font=c(title=2,leg=1,note=3),cex=c(title=1.5,leg=1,note=.7),sud=TRUE,labels=TRUE,
+  labels.filter='_|\\.',labels.trim=20,points=TRUE,points.first=TRUE,byx=TRUE,drop=c(x=TRUE,by=TRUE,bet=TRUE),
+  prat=c(1,1),model=FALSE,options=NULL,add=NULL){
   #parsing input and preparing data
-  if(!missing(options)){
+  if(!missing(options) && is.list(options) && length(options)!=0){
     a=as.list(match.call())[-1]
     options=tryCatch(options,error=function(e)NULL)
     if(is.null(options)) stop('could not find options')
@@ -353,13 +357,14 @@ splot=function(y,x=NULL,by=NULL,between=NULL,cov=NULL,type='',split='median',dat
   if(length(txt$bet)>2) txt$bet=txt$bet[1:2]
   tdc=function(x,l=NULL){
     if(is.character(x)) x=parse(text=x)
-    tx=tryCatch(eval(x,data,globalenv()),error=function(e)NULL)
+    tx=tryCatch(eval(x,data,parent.frame(2)),error=function(e)NULL)
     if(is.character(tx) && length(tx)<2){
       x=parse(text=tx)
-      tx=tryCatch(eval(x,data,globalenv()),error=function(e)NULL)
-    }
+      tx=tryCatch(eval(x,data,parent.frame(2)),error=function(e)NULL)
+    }else if(is.null(tx)) tx=tryCatch(eval(x,data,parent.frame(3)),error=function(e)NULL)
     if(is.null(tx)) stop('could not find ',x,call.=FALSE)
-    if(!is.null(l)) if(length(tx)!=l) warning(x,' is not the same length as y',call.=FALSE)
+    if(!is.null(l) && is.null(ncol(tx))) if(length(tx)!=l)
+      warning(x,' is not the same length as y',call.=FALSE)
     tx
   }
   dat=data.frame(y=tdc(txt$y))
@@ -377,12 +382,15 @@ splot=function(y,x=NULL,by=NULL,between=NULL,cov=NULL,type='',split='median',dat
     dat$cov=cbind(dat$cov,dat$x[,-1])
     dat$x=dat$x[,1]
   }
-  rm(data)
   dat=na.omit(dat)
   if(nrow(dat)==0) stop('this combination of variables/splits has no complete cases')
   dn=colnames(dat)
   if(sum(grepl('^y',dn))>1){
     #setting up multiple y variables
+    if(!ck$d){
+      tcn=if(is.null(names(tcn<-tdc(txt$y)))) colnames(tcn) else names(tcn)
+      if(!is.null(tcn) && length(tcn)==ncol(dat)) colnames(dat)=tcn
+    }
     ck$mv=TRUE
     if(ck$mlvn) lvn=FALSE
     if(!missing(by)){
@@ -580,18 +588,12 @@ splot=function(y,x=NULL,by=NULL,between=NULL,cov=NULL,type='',split='median',dat
   if(is.character(labx)) ptxt$x=labx
   if(is.character(laby)) ptxt$y=laby
   #figuring out parts of the plot
-  cs=FALSE
-  if(ck$co || (cs<-(length(colors)==1 && grepl('^grey|^gra|^past|^prim|^bright|^dark',colors,TRUE)))){
-    colors=if((ck$co && seg$by$ll>1 && seg$by$ll<9) || (cs && !grepl('^grey|^gra',colors,TRUE))){
-      if(cs && grepl('^prim|^bright',colors,TRUE)){
-        c('#45ff00','#ba00ff','#000000','#ff0000','#fffd00','#003dff','#00f2f8','#999999','#ff891b')
-      }else if(cs && grepl('^dark',colors,TRUE)){
-        c('#1b8621','#681686','#2a2a2a','#7c0d0d','#b5bc00','#241c80','#1a7e8b','#666666','#b06622')
-      }else c('#82c473','#a378c0','#616161','#9f5c61','#d3d280','#6970b2','#78c4c2','#454744','#d98c82')
-    }else if(seg$by$ll>1) grey(.2:seg$by$ll/(seg$by$ll+seg$by$ll*ifelse(seg$by$ll<10,.1,.3))) else '#999999'
-  }
+  if(!missing(colors)){
+    colors=eval(substitute(colors),data)
+    if(length(colors)==1 && grepl('^bri|^dar|^pas|^gra|^grey',colors,TRUE)) colors=splot.color(colors)
+  }else colors=if(seg$by$ll>1 && seg$by$ll<9) splot.color(colors) else splot.color('grey',ns=seg$by$ll)
   cs=colors
-  colors=colors[seq_len(seg$by$ll)]
+  colors=rep_len(colors,seg$by$ll)
   if(lvn) ptxt$l.by=paste0(paste0(ptxt$by,': '),ptxt$l.by)
   names(colors)=names(ptxt$l.by)=seg$by$l
   if(length(colors)==1) colors[2]=if(!ck$co && length(cs)>1) cs[2] else if(ck$t==3) '#999999' else '#adadad'
@@ -619,7 +621,7 @@ splot=function(y,x=NULL,by=NULL,between=NULL,cov=NULL,type='',split='median',dat
   seg$dim=if(any(ckl<-c('mfrow','mfcol')%in%names(pdo))) pdo[[if(ckl[1]) 'mfrow' else 'mfcol']] else
     if(!seg$f1$e) c(1,1) else if(!seg$f2$e){
       if(seg$f1$ll>2) l2m(seg$f1$ll) else c(2,1)
-  }else c(seg$f1$ll,seg$f2$ll)
+    }else c(seg$f1$ll,seg$f2$ll)
   seg$l=t(data.frame(strsplit(names(cdat),'^^',fixed=TRUE)))
   if(seg$f1$e){
     rownames(seg$l)=match(seg$l[,1],seg$f1$l)
@@ -785,8 +787,8 @@ splot=function(y,x=NULL,by=NULL,between=NULL,cov=NULL,type='',split='median',dat
         drop['x']=FALSE
         line.type='b'
       }
-      dx=apply(is.na(re$m),2,all)
-      if(drop['x']) re=lapply(re,function(s)s[,!dx,drop=FALSE])
+      dx=!apply(is.na(re$m),2,all)
+      if(drop['x']) re=lapply(re,function(s)s[,dx,drop=FALSE])
       m=re$m
       ne=re$ne
       pe=re$pe
@@ -808,7 +810,7 @@ splot=function(y,x=NULL,by=NULL,between=NULL,cov=NULL,type='',split='median',dat
       colnames(m)=if(drop['x'] && sum(dx)==ncol(m)) ptxt$l.x[dx] else ptxt$l.x
       stw=strwidth(colnames(m),'i')
       if((missing(xlas) || xlas>1) && sum(stw)>
-        par('fin')[1]-sum(par('omi')[c(2,4)])-dm[2]*.1 && par('fin')[1]>2.5){
+          par('fin')[1]-sum(par('omi')[c(2,4)])-dm[2]*.1 && par('fin')[1]>2.5){
         xlas=3
         if(missing(mxl)) mxl=c(1,dm[2])
         mh=c(par('fin')[2]/2,max(stw))
@@ -929,7 +931,7 @@ splot=function(y,x=NULL,by=NULL,between=NULL,cov=NULL,type='',split='median',dat
               if(grepl('^e|^co|^d',lines,TRUE)) 'e' else 'li'
           fit=tryCatch({
             if(ck$c) lm(y~x+as.matrix(td[,cvar,drop=FALSE]))$fitted else
-            if(lines=='e') y else predict(switch(lines,li=lm,lo=loess,sm=smooth.spline)(y~x))
+              if(lines=='e') y else predict(switch(lines,li=lm,lo=loess,sm=smooth.spline)(y~x))
           },error=function(e){warning('error estimating line: ',e$message,call.=FALSE);NULL})
           if(!is.null(fit)){
             if(lines=='sm') {xo=fit$x; fit=fit$y} else {or=order(x); xo=x[or]; fit=fit[or]}